Initial user service has been removed.
UserService project is now using refactored user service
