+++ RedNimbus/UserService/Program.cs
@@ -15,13 +15,12 @@ namespace RedNimbus.UserService
             SimpleIoc.Default.Register<ITokenManager, TokenManager.TokenManager>();
             SimpleIoc.Default.Register<UserService>();
             
-            SimpleIoc.Default.Register<UserService2.UserService2>();
+            SimpleIoc.Default.Register<UserService>();
             SimpleIoc.Default.Register<IUserCommunicationService,UserCommunicationService>();
 
             var userService = SimpleIoc.Default.GetInstance<UserService>();
-            var userService2 = SimpleIoc.Default.GetInstance<UserService2.UserService2>();
 
-            userService2.Start();
+            userService.Start();
         }
     }
 }
