+++ RedNimbus/UserService/UserService.cs
@@ -1,235 +1,138 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IdentityModel.Tokens.Jwt;
-using System.Text;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
+﻿using Microsoft.EntityFrameworkCore;
 using NetMQ;
 using RedNimbus.Communication;
 using RedNimbus.Domain;
 using RedNimbus.Messages;
 using RedNimbus.TokenManager;
 using RedNimbus.UserService.Helper;
+using System;
 using UserService.Database;
 using ErrorCode = RedNimbus.Either.Enums.ErrorCode;
 
 namespace RedNimbus.UserService
 {
-    public class UserService : BaseService
+    class UserService
     {
-        private static readonly Dictionary<string, string> tokenEmailPairs = new Dictionary<string, string>();
-        private IUserRepository _userRepository;
-        private ITokenManager _tokenManager;
+        private IUserRepository             _userRepository;
+        private ITokenManager               _tokenManager;
+        private IUserCommunicationService                           _userCommunicationService;
 
-        public UserService(IUserRepository repository, ITokenManager tokenManager) : base()
+        public UserService(IUserRepository userRepository, ITokenManager tokenManager, IUserCommunicationService userCommunicationService)
         {
-            Subscribe("RegisterUser", HandleRegisterUser);
-            Subscribe("AuthenticateUser", HandleAuthenticateUser);
-            Subscribe("GetUser", HandleGetUser);
-            Subscribe("DeactivateUserAccount", HandleDeactivateUserAccount);
-            
-            _tokenManager = tokenManager;
-            _userRepository = repository;
+            this._userRepository            = userRepository;
+            this._userCommunicationService  = userCommunicationService;
+            this._tokenManager              = tokenManager;
+            this.SubscribeToTopics();
         }
 
-
-
-        private bool Validate(Message<UserMessage> userMessage)
+        private void SubscribeToTopics()
         {
-            if (!Validation.IsFirstNameValid(userMessage.Data.FirstName))
-            {
-                SendErrorMessage("First Name is empty.", ErrorCode.FirstNameNullEmptyOrWhiteSpace, userMessage.Id);
-                return false;
-            }
-
-            if (!Validation.IsLastNameValid(userMessage.Data.LastName))
-            {
-                SendErrorMessage("Last Name is empty.", ErrorCode.LastNameNullEmptyOrWhiteSpace, userMessage.Id);
-                return false;
-            }
-
-            if (!Validation.IsEmailValid(userMessage.Data.Email))
-            {
-                SendErrorMessage("Invalid email format.", ErrorCode.EmailWrongFormat, userMessage.Id);
-                return false;
-            }
-
-            if (!Validation.IsPasswordValid(userMessage.Data.Password))
-            {
-                SendErrorMessage("Password does not satisfy requirements.", ErrorCode.PasswordWrongFormat, userMessage.Id);
-                return false;
-            }
-            if (!String.IsNullOrWhiteSpace(userMessage.Data.PhoneNumber) && !Validation.IsPhoneValid(userMessage.Data.PhoneNumber))
-            {
-                SendErrorMessage("Phone number wrong format", ErrorCode.PhoneNumberWrongFormat, userMessage.Id);
-                return false;
-            }
-
-            return true;
+            ((BaseService) _userCommunicationService).Subscribe("RegisterUser",             HandleRegisterUser);
+            ((BaseService) _userCommunicationService).Subscribe("AuthenticateUser",         HandleAuthenticateUser);
+            ((BaseService) _userCommunicationService).Subscribe("GetUser",                  HandleGetUser);
+            ((BaseService) _userCommunicationService).Subscribe("DeactivateUserAccount",    HandleDeactivateUserAccount);
         }
 
         private void HandleRegisterUser(NetMQMessage message)
         {
-            Message<UserMessage> userMessage = new Message<UserMessage>(message);
+            Message<UserMessage> userMessage = _userCommunicationService.HandleRegisterUserRequest(message);
 
-            if(!Validate(userMessage))
+            if(userMessage == null)
             {
                 return;
             }
 
-            User user = new User
+            User newUser = new User
             {
-                Id = Guid.NewGuid(),
-                FirstName = userMessage.Data.FirstName,
-                LastName = userMessage.Data.LastName,
-                Email = userMessage.Data.Email,
-                Password = HashHelper.ComputeHash(userMessage.Data.Password),
+                Id          = Guid.NewGuid(),
+                FirstName   = userMessage.Data.FirstName,
+                LastName    = userMessage.Data.LastName,
+                Email       = userMessage.Data.Email,
+                Password    = HashHelper.ComputeHash(userMessage.Data.Password),
                 PhoneNumber = userMessage.Data.PhoneNumber,
             };
 
             try
             {
-                this._userRepository.SaveUser(user);
-
-                userMessage.Topic = "Response";
-
-                NetMQMessage msg = userMessage.ToNetMQMessage();
-                SendMessage(msg);
+                _userRepository.SaveUser(newUser);
+                _userCommunicationService.HandleRegisterUserResponse(userMessage, newUser);
             }
             catch (DbUpdateException)
             {
-                SendErrorMessage("Email already exists.", ErrorCode.EmailAlreadyUsed, userMessage.Id);
+                _userCommunicationService.SendUserErrorMessage("Email already exists", ErrorCode.EmailAlreadyUsed, userMessage.Id);
             }
             catch (Exception)
             {
-                SendErrorMessage("Internal server error.", ErrorCode.InternalServerError, userMessage.Id);
+                _userCommunicationService.SendUserErrorMessage("Internal  server error", ErrorCode.InternalServerError, userMessage.Id);
             }
         }
 
         private void HandleAuthenticateUser(NetMQMessage message)
         {
-            Message<UserMessage> userMessage = new Message<UserMessage>(message);
-
-            if (!Validation.IsEmailValid(userMessage.Data.Email))
-            {
-                SendErrorMessage("Email or password are not valid!", ErrorCode.IncorrectEmailOrPassword, userMessage.Id);
-            }
-
-            if (!Validation.IsPasswordValid(userMessage.Data.Password))
-            {
-                SendErrorMessage("Email or password are not valid!", ErrorCode.IncorrectEmailOrPassword, userMessage.Id);
-            }
+            Message<UserMessage> userMessage = _userCommunicationService.HandleAuthenticateUserRequest(message);
 
             var email = userMessage.Data.Email;
+
             if (_userRepository.CheckIfExists(email))
-                {
+            {
                 var registeredUser = _userRepository.GetUserByEmail(userMessage.Data.Email);
-                if (registeredUser.Password == HashHelper.ComputeHash(userMessage.Data.Password))
+                if(registeredUser.Password == HashHelper.ComputeHash(userMessage.Data.Password))
                 {
-                    Message<TokenMessage> tokenMessage = new Message<TokenMessage>("Response");
-
-                    tokenMessage.Id = userMessage.Id;
-                    tokenMessage.Data.Token = _tokenManager.GenerateToken(registeredUser.Id);
-
-
-                    if (!tokenEmailPairs.ContainsKey(tokenMessage.Data.Token))
-                    {
-                        tokenEmailPairs.Add(tokenMessage.Data.Token, userMessage.Data.Email);
-                    }
-
-                    NetMQMessage msg = tokenMessage.ToNetMQMessage();
-                    SendMessage(msg);
-                    return;
+                    var token = _tokenManager.GenerateToken(registeredUser.Id);
+                    _userCommunicationService.HandleAuthenticateUserResponse(userMessage, token);
                 }
             }
-            SendErrorMessage("Email or password are not valid!", ErrorCode.IncorrectEmailOrPassword, userMessage.Id);
+            _userCommunicationService.SendUserErrorMessage("Email or password are not valid", ErrorCode.IncorrectEmailOrPassword, userMessage.Id);
         }
-
-
 
         private void HandleGetUser(NetMQMessage message)
         {
-            Message<TokenMessage> tokenMessage = new Message<TokenMessage>(message);
-
-            if (tokenMessage.Data.Token == null)
-            {
-                SendErrorMessage("Requested user data not found", ErrorCode.UserNotFound, tokenMessage.Id);
-                return;
-            }
+            Message<TokenMessage> tokenMessage = _userCommunicationService.HandleGetUserRequest(message);
 
             Guid id = _tokenManager.ValidateToken(tokenMessage.Data.Token);
+
             if (id.Equals(Guid.Empty))
             {
-                SendErrorMessage("Requested user data not found", ErrorCode.UserNotFound, tokenMessage.Id);
+                _userCommunicationService.SendUserErrorMessage("Requested user data not found", ErrorCode.UserNotFound, tokenMessage.Id);
                 return;
             }
 
             User registeredUser = _userRepository.GetUserById(id);
             if (registeredUser == null)
             {
-                SendErrorMessage("Requested user data not found", ErrorCode.UserNotRegistrated, tokenMessage.Id);
+                _userCommunicationService.SendUserErrorMessage("Requested user data not found", ErrorCode.UserNotRegistrated, tokenMessage.Id);
                 return;
             }
 
-            Message<UserMessage> userMessage = new Message<UserMessage>("Response")
-            {
-                Id = tokenMessage.Id,
-                Data = new UserMessage
-                {
-                    FirstName = registeredUser.FirstName,
-                    LastName = registeredUser.LastName,
-                    Token = tokenMessage.Data.Token
-                }
-            };
-
-            NetMQMessage msg = userMessage.ToNetMQMessage();
-            SendMessage(msg);
+            _userCommunicationService.HandleGetUserResponse(tokenMessage, registeredUser);
         }
 
         private void HandleDeactivateUserAccount(NetMQMessage message)
         {
-            Message<TokenMessage> tokenMessage = new Message<TokenMessage>(message);
-
-            if (tokenMessage.Data.Token == null)
-            {
-                SendErrorMessage("Requested user data not found", ErrorCode.UserNotFound, tokenMessage.Id);
-                return;
-            }
+           Message<TokenMessage> tokenMessage = _userCommunicationService.HandleDeactivateUserAccountRequest(message);
 
             Guid id = _tokenManager.ValidateToken(tokenMessage.Data.Token);
+
             if (id.Equals(Guid.Empty))
             {
-                SendErrorMessage("Requested user data not found", ErrorCode.UserNotFound, tokenMessage.Id);
+                _userCommunicationService.SendUserErrorMessage("Requested user data not found", ErrorCode.UserNotFound, tokenMessage.Id);
                 return;
             }
 
             if (_userRepository.GetUserById(id) == null)
             {
-                SendErrorMessage("Requested user data not found", ErrorCode.UserNotRegistrated, tokenMessage.Id);
+                _userCommunicationService.SendUserErrorMessage("Requested user data not found", ErrorCode.UserNotRegistrated, tokenMessage.Id);
                 return;
             }
 
             _userRepository.DeactivateUserAccount(id);
 
-            tokenMessage.Topic = "Response";
-            NetMQMessage msg = tokenMessage.ToNetMQMessage();
-            SendMessage(msg);
+            _userCommunicationService.HandleDeactivateUserAccountResponse(tokenMessage);
         }
 
-        private void SendErrorMessage(string messageText, ErrorCode errorCode, NetMQFrame idFrame)
+        public void Start()
         {
-            Message<ErrorMessage> errorMessage = new Message<ErrorMessage>("Error")
-            {
-                Data = new ErrorMessage
-                {
-                    MessageText = messageText,
-                    ErrorCode = (int)errorCode
-                },
-                Id = idFrame
-            };
-
-            NetMQMessage msg = errorMessage.ToNetMQMessage();
-            SendMessage(msg);
+           ((BaseService)_userCommunicationService).Start();
         }
     }
 }
